+++ lib/attesto_phoenix/controller/token_controller.ex
@@ -194,7 +194,8 @@ defmodule AttestoPhoenix.Controller.TokenController do
     policy = %Policy{
       allow_public: true,
       assertion_audiences: [config.issuer, Config.token_endpoint_url(config)],
-      assertion_max_lifetime: @client_assertion_max_lifetime
+      assertion_max_lifetime: @client_assertion_max_lifetime,
+      assertion_signing_algs: config.client_auth_signing_algs
     }
 
     case ClientAuthentication.authenticate(